cos: split the #6846 resolver cells out of cos.rs
The modularity gate fired: forwarding_build/cos.rs went 1036 -> 1530 LOC
and crossed the 1500 WATCH floor. Split rather than recorded as an
exception in docs/refactoring-audit-accepted.txt — the file had grown a
~300 line inline `#[cfg(test)]` module, which is both the easiest 300
lines to move and the ones with the fewest ties to the rest of the file.

Moved to forwarding_build/cos/remainder_temporal_tests_6846.rs, loaded as
a CHILD module of `cos` rather than a `#[path]` sibling, so `use super::*`
still reaches the private resolver functions these cells exist to pin.
The module path is unchanged, so `cargo test
remainder_temporal_tests_6846` still selects all 13 cells rather than
silently selecting zero.

The builder-level cells stay in forwarding_build/tests.rs alongside the
other build_cos_state cells. That separation is the point of the F5
finding: one file pins what the resolver RETURNS, the other pins what a
queue ENDS UP WITH, and it was the absence of the second that let an
argument swap at the call site pass a green crate.

cos.rs is 1235 LOC after the split.

Validation: go test ./pkg/refactoraudit/ -count=1 is ok; cargo test
--bin xpf-userspace-dp remainder_temporal -- --test-threads=1 runs all
13 cells at the same module path; rustfmt --check on the new file is
clean.
