+++ docs/log/6846.md
@@ -157,3 +157,21 @@
     `pkg/config/schema_cos_buffer_temporal_4228_test.go`,
     `userspace-dp/src/afxdp/forwarding_build/tests.rs`,
     `docs/cos-traffic-shaping.md`, `docs/log/6846.md`
+
+- **Timestamp**: 2026-08-27
+- **Action**: Split the #6846 resolver cells out of `forwarding_build/cos.rs`
+  - The modularity gate fired: `cos.rs` went 1036 -> 1530 LOC and crossed the
+    1500 WATCH floor. Split rather than recorded as an exception — the file had
+    grown a ~300 line inline `#[cfg(test)]` module, which is the easiest 300
+    lines in it to move and the ones with the fewest ties to the rest.
+  - Moved to `forwarding_build/cos/remainder_temporal_tests_6846.rs`, loaded as
+    a CHILD module of `cos` (not a `#[path]` sibling), so `use super::*` still
+    reaches the private resolver functions. The module path is unchanged, so a
+    `cargo test remainder_temporal_tests_6846` filter still selects all 13.
+  - The builder-level cells stay in `forwarding_build/tests.rs` with the other
+    `build_cos_state` cells. That separation is the point of F5: one file pins
+    what the resolver RETURNS, the other pins what a queue ENDS UP WITH.
+  - `cos.rs` is 1235 LOC after the split; the gate is green.
+  - **File(s)**: `userspace-dp/src/afxdp/forwarding_build/cos.rs`,
+    `userspace-dp/src/afxdp/forwarding_build/cos/remainder_temporal_tests_6846.rs`,
+    `docs/log/6846.md`
